Applied requested changes
Set the ID as optional, since it is generated automatically when adding an expense to the database.
Add a condition in edit-dialog.component.html to determine the section title.
I kept only one instance for each key, ensuring that the values are set correctly based on the existence of this.data.expense.